refactor: centralize time series types
